fix: resolve 22 critical/high/medium bugs across the codebase
Critical fixes:
- Add missing logger in ml/tasks.py (NameError in notification helpers)
- Fix duplicate delete_model method in model_service (hard delete overwrote soft delete)
- Fix restore_dataset unable to find archived datasets (get_dataset now supports include_archived)
- Fix ModelStatus.PRODUCTION AttributeError in retention_service (use DEPLOYED)
- Remove duplicate ModelShare model from advanced.py (table name conflict)
- Fix duplicate AuditLog in models/audit.py + mutable default={} data corruption
- Add require_admin checks to marketplace review/moderate endpoints
- Add try/except for Content-Length parsing in security middleware

High fixes:
- Fix authz.py ownership bypass when owner_id is None (IDOR vulnerability)
- Fix Redis connection leak on failed ping
- Fix ZeroDivisionError in automl_task with empty algorithms list
- Fix automl_task returning success after DB rollback + add model artifact saving
- Fix SQL double comma syntax error in scraping_service batch/recursive queries
- Fix TOCTOU race condition in serving_service pipeline cache (double-check inside lock)
- Fix TOCTOU race condition in api_quota check_and_increment (SELECT FOR UPDATE)
- Fix mass assignment vulnerability in update_model (field whitelist)
- Fix retention service file deletion before DB flush

Medium fixes:
- Remove unsafe-inline from CSP script-src header
- Fix feature_store_service None unpack on snapshot.features
- Fix experiment.duration_seconds String -> Float column type
- Fix ABTest accuracy columns Integer -> Float
- Fix DataQualityReport count columns Float -> Integer
- Add ownership checks to lineage_metrics endpoints
- Fix feature_monitoring sync blocking call with asyncio.to_thread
- Fix external_data premature db.commit() breaking transactions
- Fix quota admin setting tier on wrong user

+++ app/models/ab_test.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime, timezone
-from sqlalchemy import Column, String, DateTime, ForeignKey, JSON, Integer, Boolean, Enum, Index
+from sqlalchemy import Column, String, DateTime, ForeignKey, JSON, Integer, Boolean, Enum, Index, Float
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 import enum
@@ -28,8 +28,8 @@ class ABTest(Base):
     model_b_id = Column(UUID(as_uuid=True), ForeignKey("models.id"), nullable=False)
     model_a_requests = Column(Integer, default=0)
     model_b_requests = Column(Integer, default=0)
-    model_a_accuracy = Column(Integer, default=0)
-    model_b_accuracy = Column(Integer, default=0)
+    model_a_accuracy = Column(Float, default=0.0)
+    model_b_accuracy = Column(Float, default=0.0)
     results = Column(JSON, default=dict)
     started_at = Column(DateTime)
     ended_at = Column(DateTime)
